refactor: split image source extraction leaf modules
